+++ plugin/src/models/gatsby-api.ts
@@ -2,55 +2,98 @@ import { GatsbyHelpers } from "~/utils/gatsby-types"
 import merge from "lodash/merge"
 import { createRemoteMediaItemNode } from "~/steps/source-nodes/create-nodes/create-remote-media-item-node"
 import { menuBeforeChangeNode } from "~/steps/source-nodes/before-change-node/menu"
+import { cloneDeep } from "lodash"
 
+export interface PluginOptionsPreset {
+  presetName: string
+  useIf: (helpers: GatsbyHelpers, pluginOptions: IPluginOptions) => boolean
+  options: IPluginOptions
+}
+
+const inDevelopPreview =
+  process.env.NODE_ENV === `development` &&
+  !!process.env.ENABLE_GATSBY_REFRESH_ENDPOINT
+
+const inPreviewRunner = process.env.RUNNER_TYPE === `PREVIEW`
+
+export const previewOptimizationPreset: PluginOptionsPreset = {
+  presetName: `PREVIEW_OPTIMIZATION`,
+  useIf: (): boolean => inDevelopPreview || inPreviewRunner,
+  options: {
+    html: {
+      useGatsbyImage: false,
+      createStaticFiles: false,
+    },
+    type: {
+      __all: {
+        limit: 50,
+      },
+      Comment: {
+        limit: 0,
+      },
+      Menu: {
+        limit: null,
+      },
+      MenuItem: {
+        limit: null,
+      },
+      User: {
+        limit: null,
+      },
+    },
+  },
+}
 export interface IPluginOptions {
-  url: string
-  verbose: boolean
-  debug: {
-    throwRefetchErrors: boolean
-    graphql: {
-      showQueryOnError: boolean
-      showQueryVarsOnError: boolean
-      copyQueryOnError: boolean
-      panicOnError: boolean
-      onlyReportCriticalErrors: boolean
-      copyNodeSourcingQueryAndExit: boolean
-      writeQueriesToDisk: boolean
+  url?: string
+  verbose?: boolean
+  debug?: {
+    throwRefetchErrors?: boolean
+    graphql?: {
+      showQueryOnError?: boolean
+      showQueryVarsOnError?: boolean
+      copyQueryOnError?: boolean
+      panicOnError?: boolean
+      onlyReportCriticalErrors?: boolean
+      copyNodeSourcingQueryAndExit?: boolean
+      writeQueriesToDisk?: boolean
     }
-    timeBuildSteps: boolean
-    disableCompatibilityCheck: boolean
+    timeBuildSteps?: boolean
+    disableCompatibilityCheck?: boolean
   }
-  develop: {
-    nodeUpdateInterval: number
-    hardCacheMediaFiles: boolean
-    hardCacheData: boolean
+  develop?: {
+    nodeUpdateInterval?: number
+    hardCacheMediaFiles?: boolean
+    hardCacheData?: boolean
   }
-  production: {
-    hardCacheMediaFiles: boolean
+  production?: {
+    hardCacheMediaFiles?: boolean
   }
-  auth: {
+  auth?: {
     htaccess: {
       username: string | null
       password: string | null
     }
   }
-  schema: {
+  schema?: {
     queryDepth: number
     circularQueryLimit: number
     typePrefix: string
     timeout: number // 30 seconds
     perPage: number
+    requestConcurrency?: number
   }
-  excludeFieldNames: []
-  html: {
-    useGatsbyImage: boolean
-    imageMaxWidth: number
-    fallbackImageMaxWidth: number
-    imageQuality: number
-    createStaticFiles: boolean
+  excludeFieldNames?: []
+  html?: {
+    useGatsbyImage?: boolean
+    imageMaxWidth?: number
+    fallbackImageMaxWidth?: number
+    imageQuality?: number
+    createStaticFiles?: boolean
   }
-  type: {
+  presets?: PluginOptionsPreset[]
+  type?: {
     [typename: string]: {
+      limit?: number
       excludeFieldNames?: string[]
       exclude?: boolean
       // @todo type this
@@ -61,6 +104,7 @@ export interface IPluginOptions {
       localFile?: {
         excludeByMimeTypes?: string[]
         maxFileSizeBytes?: number
+        requestConcurrency?: number
       }
     }
   }
@@ -103,6 +147,7 @@ const defaultPluginOptions: IPluginOptions = {
     typePrefix: `Wp`,
     timeout: 30 * 1000, // 30 seconds
     perPage: 100,
+    requestConcurrency: 15,
   },
   excludeFieldNames: [],
   html: {
@@ -121,6 +166,7 @@ const defaultPluginOptions: IPluginOptions = {
     // Also fetches those files if they don't already exist
     createStaticFiles: true,
   },
+  presets: [previewOptimizationPreset],
   type: {
     __all: {
       // @todo make dateFields into a plugin option?? It's not currently
@@ -158,6 +204,7 @@ const defaultPluginOptions: IPluginOptions = {
       localFile: {
         excludeByMimeTypes: [],
         maxFileSizeBytes: 15728640, // 15Mb
+        requestConcurrency: 100,
       },
       beforeChangeNode: async ({
         remoteNode,
@@ -289,6 +336,7 @@ const defaultPluginOptions: IPluginOptions = {
 interface IGatsbyApiState {
   helpers: GatsbyHelpers
   pluginOptions: IPluginOptions
+  activePluginOptionsPresets: PluginOptionsPreset[]
 }
 
 const gatsbyApi = {
@@ -302,7 +350,39 @@ const gatsbyApi = {
       state: IGatsbyApiState,
       payload: IGatsbyApiState
     ): IGatsbyApiState {
-      return merge(state, payload)
+      const stateCopy = cloneDeep(state)
+
+      const defaultPresets = stateCopy.pluginOptions?.presets || []
+      const userPresets = payload.pluginOptions?.presets || []
+
+      /**
+       * Presets are plugin option configurations that are conditionally
+       * applied based on a `useIf` function (which returns a boolean)
+       * If it returns true, that preset is used.
+       */
+      const optionsPresets = [
+        ...defaultPresets,
+        ...userPresets,
+      ]?.filter((preset) =>
+        preset.useIf(payload.helpers, payload.pluginOptions)
+      )
+
+      if (optionsPresets?.length) {
+        state.activePluginOptionsPresets = optionsPresets
+
+        let presetModifiedOptions = state.pluginOptions
+
+        for (const preset of optionsPresets) {
+          presetModifiedOptions = merge(presetModifiedOptions, preset.options)
+        }
+
+        state.pluginOptions = presetModifiedOptions
+      }
+
+      // add the user defined plugin options last so they override any presets
+      state = merge(state, payload)
+
+      return state
     },
   },
 }